backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
snowflake-connector-python==3.15.0
sentence-transformers==3.3.1
scikit-learn==1.6.0
pandas==2.2.3
numpy==2.2.6
python-dotenv==1.0.1
pydantic==2.10.4
python-multipart==0.0.20
anthropic==0.40.0

# PDF Processing
PyMuPDF==1.24.14
pypdf2==3.0.1

# File handling and validation
aiofiles==24.1.0
python-magic==0.4.27

# Background job processing
celery==5.4.0
redis==5.2.1

# Additional utilities
uuid==1.30
jsonschema==4.23.0
flask==3.1.0
flask-cors==5.0.0
